+++ CHANGELOG.md
@@ -36,3 +36,5 @@ December 5th, 2023
     - Predator's real-time ALPR functionality can now be registered as a SystemD service from the Cortex interface.
     - Cortex now allows users to view the contents of files in the interface, instance, and working directories.
     - Added tool to view and download the plate history log in various formats.
+- It is now possible to configure Cortext to not display each guess for each plate.
+- Cortex now displays plates involved in alerts with a red background.